backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==12.0
python-multipart==0.0.9
pydantic-settings==2.5.0
adversarial-robustness-toolbox==1.18.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-image>=0.21.0
scipy>=1.11.0
